+++ packages/sdk-js/src/GrowthBook.ts
@@ -113,7 +113,11 @@ export class GrowthBook<
 
   public async loadFeatures(options?: LoadFeaturesOptions): Promise<void> {
     await this._refresh(options, true, true);
-    if (options && options.autoRefresh) {
+
+    if (
+      this._ctx.backgroundSync !== false &&
+      (this._ctx.subscribeToChanges || (options && options.autoRefresh))
+    ) {
       subscribe(this);
     }
   }
@@ -146,7 +150,7 @@ export class GrowthBook<
       options.skipCache || this._ctx.enableDevMode,
       allowStale,
       updateInstance,
-      this._ctx.enableStreaming !== false
+      this._ctx.backgroundSync !== false
     );
   }