+++ packages/sdk-js/src/types/growthbook.ts
@@ -149,7 +149,8 @@ export interface Context {
   forcedVariations?: Record<string, number>;
   log?: (msg: string, ctx: any) => void;
   qaMode?: boolean;
-  enableStreaming?: boolean;
+  backgroundSync?: boolean;
+  subscribeToChanges?: boolean;
   enableDevMode?: boolean;
   /* @deprecated */
   disableDevTools?: boolean;